requirements.txt
python-chess
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
tqdm
requests
flask
pytest

